refactoring and setting up weapon rack renderer
