+++ src/main/java/com/onewhohears/dscombat/client/event/ClientModEvents.java
@@ -7,13 +7,13 @@ import com.onewhohears.dscombat.client.renderer.RendererEntityAbstractAircraft;
 import com.onewhohears.dscombat.client.renderer.RendererEntityAbstractWeapon;
 import com.onewhohears.dscombat.client.renderer.RendererEntityInvisible;
 import com.onewhohears.dscombat.client.renderer.RendererEntityWeaponRack;
-import com.onewhohears.dscombat.client.renderer.model.EntityModelAlexisPlane;
-import com.onewhohears.dscombat.client.renderer.model.EntityModelBullet1;
-import com.onewhohears.dscombat.client.renderer.model.EntityModelF16;
-import com.onewhohears.dscombat.client.renderer.model.EntityModelJaviPlane;
-import com.onewhohears.dscombat.client.renderer.model.EntityModelMissile1;
-import com.onewhohears.dscombat.client.renderer.model.EntityModelNoahChopper;
-import com.onewhohears.dscombat.client.renderer.model.EntityModelTestPlane;
+import com.onewhohears.dscombat.client.renderer.model.aircraft.EntityModelAlexisPlane;
+import com.onewhohears.dscombat.client.renderer.model.aircraft.EntityModelF16;
+import com.onewhohears.dscombat.client.renderer.model.aircraft.EntityModelJaviPlane;
+import com.onewhohears.dscombat.client.renderer.model.aircraft.EntityModelNoahChopper;
+import com.onewhohears.dscombat.client.renderer.model.aircraft.EntityModelTestPlane;
+import com.onewhohears.dscombat.client.renderer.model.weapon.EntityModelBullet1;
+import com.onewhohears.dscombat.client.renderer.model.weapon.EntityModelMissile1;
 import com.onewhohears.dscombat.entity.aircraft.EntityHelicopter;
 import com.onewhohears.dscombat.entity.aircraft.EntityPlane;
 import com.onewhohears.dscombat.entity.parts.EntityWeaponRack;
