+++ src/main/java/com/onewhohears/dscombat/client/renderer/RendererEntityWeaponRack.java
@@ -3,20 +3,26 @@ package com.onewhohears.dscombat.client.renderer;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.onewhohears.dscombat.entity.parts.EntityWeaponRack;
 
+import net.minecraft.client.model.EntityModel;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.entity.EntityRenderer;
 import net.minecraft.client.renderer.entity.EntityRendererProvider.Context;
 import net.minecraft.resources.ResourceLocation;
 
 public class RendererEntityWeaponRack<T extends EntityWeaponRack> extends EntityRenderer<T> {
-
-	public RendererEntityWeaponRack(Context context) {
+	
+	protected final EntityModel<?> model;
+	protected final ResourceLocation texture;
+	
+	public RendererEntityWeaponRack(Context context, EntityModel<?> model, ResourceLocation texture) {
 		super(context);
+		this.model = model;
+		this.texture = texture;
 	}
 
 	@Override
 	public ResourceLocation getTextureLocation(T entity) {
-		return null;
+		return texture;
 	}
 	
 	@Override
