+++ src/main/java/com/onewhohears/dscombat/client/renderer/model/aircraft/EntityModelTestPlane.java
@@ -1,4 +1,4 @@
-package com.onewhohears.dscombat.client.renderer.model;
+package com.onewhohears.dscombat.client.renderer.model.aircraft;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
