+++ src/main/java/com/onewhohears/dscombat/entity/parts/EntityWeaponRack.java
@@ -1,5 +1,7 @@
 package com.onewhohears.dscombat.entity.parts;
 
+import com.onewhohears.dscombat.entity.aircraft.EntityAbstractAircraft;
+
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.phys.Vec3;
@@ -15,5 +17,12 @@ public class EntityWeaponRack extends EntityAbstractPart {
 	}
 	
 	// TODO display weapon models under aircraft wings
+	
+	public int getMissileNum() {
+		if (this.getVehicle() instanceof EntityAbstractAircraft plane) {
+			
+		}
+		return 0;
+	}
 
 }
